feat: (http) Add automatic HTTP to HTTPS redirection
This commit adds a feature to automatically redirect HTTP requests to their HTTPS counterpart. The change is made in the `pass_through_request` function to update the URL with "https://" if it starts with "http://". This improves the security of the app and makes sure that all requests are encrypted. Additionally, a comment is added to explain why compression is disabled.